move schmitt trigger to separate device file
